fix: comprehensive bug fixes for 20-issue security/bug audit
- lexer: error reporting for integer overflow (hex/binary/octal/decimal)
- lexer: expand tabs to 4 spaces in multiline string dedent
- parser: propagate lexer errors from string interpolation sub-lexer
- typecheck: change enum tag from u8 to u32 to prevent overflow
- codegen: remove silent divide-by-zero replacement in bin_div/bin_mod
- codegen: add Stream/Task RC management in function parameter entry
- codegen: fix RC management — tree nodes RC=1, element rc_inc in push
- codegen: remove duplicate codegen rc_inc and double compilation for Map/Set/List
- codegen: add rc_free_method_receiver for method dispatch cleanup
- codegen: fix RC leak on call_rt failure in map_set handlers (#9)
- http_runtime: fix SSRF bypass with proper IP parsing (octal/hex/IPv6-mapped)
- runtime_json: fix from_cstr unconstrained lifetime (#13)
- runtime_json: fix dangling pointers in action_json_get/get_idx (#14)
- runtime_json: implement action_json_free_cstr, fix allocator mismatch (#15)
- runtime_threading: replace unsafe TerminateThread with error return on Windows (#16)
- runtime_threading: implement Windows clock_gettime via QPC/GetSystemTime (#17)
- lsp: use token-based when-keyword search instead of raw string rfind (#18)
- lsp: deduplicate completion items from type_env and definition_map (#19)
- cli: fix version to use CARGO_PKG_VERSION (#20)

+++ src/typecheck.rs
@@ -23,7 +23,7 @@ pub struct EnumInfo {
 #[derive(Clone, Debug)]
 pub struct EnumVariantInfo {
     pub name: String,
-    pub tag: u8,
+    pub tag: u32,
     pub params: Vec<EnumVariantParam>,
 }
 
@@ -77,7 +77,7 @@ impl TypeRegistry {
                     self.variant_to_enum.insert(v.name.clone(), name.clone());
                     enum_variants.push(EnumVariantInfo {
                         name: v.name.clone(),
-                        tag: i as u8,
+                        tag: i as u32,
                         params: v.params.clone(),
                     });
                 }
